Fix a potential linker problem.
The problem occurs when Boost and the compiler have the same default
installation path. In this case, CMake tries to be smart during link time and
replace the absolute path of the shared library with -l<lib>. With both
system-wide and user-provided installations, this can cause CMake to pick up
the wrong version of the library. Adding an explicit -L flag to the linker
command helps.