Reorganized configure options
Due to ongoing problems with Boost on Linux distributions, this patch disables
Boost support by default, i.e., turns context-switching off. Users can enable
it explicitly by using `--enable-context-switching`. The old configure options
to disable this feature explicitly are now deprecated. Furthermore,
`--with-opencl` is replaced by `--enable-opencl` and
`--disable-context-switching` is replaced by `--no-context-switching`. This
change implements a more consistent naming scheme. The old option names are
still valid, but give a deprecated warning to the user and will be removed in
the future.