New terminology, e.g., `any_tuple` => `message`
This patch renames several classes to give programmers without functional
background a better intuition about the purpose of a particular class. For
example, `partial_function` is not a concept familiar to all C++ developers and
has been renamed to `message_handler`. Although the former is a better and more
general name in theory, the latter is a better description for what the class
is actually used for in the library.