Fix `then` and `await`, close #143
Synchronous response message are handled properly and patterns are no longer
extended automatically. The extension caused CAF to assume `on(any_vals,
arg_match)` before any functor given to `then` or `await`. This feature is
neither documented nor is it the expected behavior from a users point of view.
This could potentially break code, but the broken code shouldn't be valid in
the first place...