Use a single queue in the worker to allow stealing
This change allows stealing at all times. It requires the worker to synchronize
its access to the queue. However, this overhead has been measured to be 4-5%
for the actor_creation benchmark which stresses the scheduler heavily by
spawning a million actors computing virtually nothing. In a "real-world"
application, the overhead is probably not even measurable.