Remove automated demangling, close #205
CAF no longer automatically tries to demangle platform-dependent type name
representations as returned by `typeid(...).name()`. Although this is a nice
convenience feature in theory, there are to many issues attached to it,
including broken ABI versions on some platforms. Instead of automatically
computing the uniform type name at runtime, users of CAF now need to pass it to
`announce`. This is a breaking change, as the signature of `announce` has been
changed to take a string as first argument.