Replace `match_expr` with `match_case` tuples
The main objective of this change is to make CAF stack traces readable,
reducing the amount of "template vomit" visible at runtime to a bare minimum.
As a side effect, this patch streamlines the implementation quite a bit to have
an easier time maintaining it in the future.

Also `others()` can now be written as simply `others`.