Remove actor policies and `proper_actor`
The previous approach based on mixins, policies, and the glue-it-all-together
`proper_actor` was ok for fast prototyping. The downside of this approach is
that stack traces are basically useless, since they are flooded with irrelevant
template metaprogramming types. Since CAF has long outgrown the prototyping
stage, it was about time to streamline actor classes. This also shortens build
times.