Report absolute instead of cumulative measurements
This has the advantage that it makes each sample independent and one can
now easily "slice" profiler log files. For example, if the first half of
the execution time is quite different from the second (due to caching or
different user input), then one can now use the first half of the log
file and the second half independently, which was not possible before
because log entries were cumulative.