Bug fixes.
After the calling of local_actor::quit, if m_pending_responses is not
empty or on_exit did set a new behavior, no matter what the exit reason
was, it will be changed to exit_reason::normal at last. So I think we
should record the exit reason at right time and resume it after the
behaviors and responses are handled.