Make message::extract_opts more flexible
This function message::extract_opts no longer prints help text and errors
to the console. These remain available in the result struct as members,
however.

Moreover, this commit also fixes a bug when users specified their own -h
or -? option, but not --help. In this case extract_opts would override
it with the default help arg. This no longer happens.

Another extension concerns handling of values for single options: it is
now possible to omit the space between a value and the single option,
e.g., -v2 is equivalent to -v 2.