Fix possible segfault during shutdown
A forwarding proxy sends a messages in its dtor to indicate unreferenced
proxies. However, after `kill_proxy` was called, the receiver no longer exists
and this can cause a segfault during shutdown if the message is send during the
dtor of the middleman. By redirecting all further messages after `kill_proxy`
to `invalid_actor`, we get rid of unnecessary messages as well as possible
segfaults.