Update syntax of spawn_cl
The new syntax describes the signature in form of a variadic number of
in, out and in_out arguments as the last arguments to spawn_cl. This
allows multiple output arguments as well as arguments used for input as
well as output. The type of the argument is passed as a template
parameter to the in, out and in_out classes.

out accepts a function to calculate the size of the result buffer
from the input arguments for a specific calculation. Per default, its
size is equal to the number of work items.

The configuration parameters for the actor (work-items, offsets, ...)
are now wrapped in a spawn_configuration.

Examples and tests have been updated accordingly. For now, the old
spawn_cl syntax is still valid, but marked as deprecated.