Fix build on MSVC
This commit includes a fix for `accept` of the socket API with MSVC.

Examples are now working, but use the `trivial match case`, the `advanced
match case` is not available on MSVC due to internal compiler errors.

The `thread_local` keyword is used on MSVC and GCC, while Clang does not
support it for now.

This commit also fixes minor warnings with the MSVC compiler and removes
the `pragma comment` for linking libraries with MSVC.