Bring back macro for checks that expect to fail
This used to be CAF_FAIL, but in the process of streamlining the
semantics of this macro to be compatible with Boost.Test, we've lost the
ability to express a check which is known to fail. The macro
CAF_CHECK_FAIL brings back this capability. The suffix _FAIL after CHECK
hints that this is a variation of CAF_CHECK.