Switch back from maybe to optional
The class `maybe` is obsoleted by `result` in all cases that could return an
error. In all other cases, `optional` is equally well suited to do the job.
Since C++17 will contain an `optional` class, it makes more sense to use this
abstraction and switch to the standard library version once available.