Write out sequences in variable byte encoding
This optimization pays is a common one in binary serialization
protocols. We lifted the implementation directly from VAST's
serialization framework. The optimization pays off when a sequence has
less than 2^21 elements, which is the common case.

Moreover, with this change we treat strings as orindary sequences.
Consequently, strings are now binary-compatible with byte vectors, in
case that ever turns out to be important.