Optimize for large contiguous byte sequences
When processing large amounts of data, it is common to store it as
std::vector<char> or std::vector<uint8_t>. Previously, such types put
heavy stress on the serialization framework, because each character was
processed separatedly. This patch changes this behavior such that the
entire sequence is processed at once.