Enable actors to receive streams as response
This patch is only one half of enabling actors to "request" streams from
others. Response IDs are currently not accounted for. This requires to
somehow signalize the *intent* of setting a message ID. The current use
case for this is starting a stream and then receiving the result.
However, when attaching oneself to a stream as sink and wanting to
provide the sink event handlers via response handler (e.g. via .then)
then actors in the pipeline need to be able to figure out which of the
two options is true.