Fix null pointer exception
As described in Issue 542, there is a null pointer exception in the inspect method.
The exception was caused by capturing a local variable as reference in a lambda and passing that lambda as a callback to another function. At the time the lambda was called, the captured references are allready gone.

Never capture local variables by ref, if you pass the lambda lives longer than the scope it is created in.