Fix data buffered in OpenSSL not being flushed
OpenSSL may report pending data (internally buffered) that is no
longer available to read from the underlying socket.  In that case,
the event loop must still be made aware to continue read operations
because otherwise the data may remain buffered in OpenSSL until the
peer sends more data on the socket -- an arbitrary amount of time,
dependent on how talkative the peer is.