Add marker to make categories optional on the CLI
Categories are great at organizing program options. However, on the CLI
they get in the way quickly. This change allows developers to prefix
category names with '?' to make it optional on the CLI. For example, an
option added via

```
opt_group{custom_options_, "?foo"}
.add<std::string>("bar,b", "some string");
```

allows users to use any of these on the CLI:

- `-b text`
- `--bar=text`
- `--foo.bar=text`

All options in the `global` category already allowed this abbreviated
form implicitly.