Don't consider uint8_t and bool as same
This extra check is necessary for instantiations of `caf::variant`s
containing `bool` and `uint8_t` members in that order. If one tries
to construct the variant with `uint8_t`, the `type_id` calculation
will place the constructor argument at the `T` argument and the
type list entry at `U`. Therefore this additional condition is
needed.