Synchronize endpoint managers only via their queue
Using both a queue plus atomic bitmask for read and write flags
introduces race conditions between the two. As a result, the endpoint
manager ends up in corrupt states. This redesign synchronizes the
endpoint manager in the same way CAF synchronizes actors: via a single
mailbox that indicates whether the manager is blocked (currently waiting
for new events or messages) or not.