Add new object-based access API
The new utility class config_value_object_access makes it much simpler
for users to specialize config_value_access. Now, users only need to
implement a trait class for config_value_object_access to get all the
low-level sum type calls as well as the parsing code out of their way.

The center piece of the trait class is a getter for fields of the
user-defined type. This new fields-based API either works with pointers
to members or with paris of getter and setter functions. In both cases,
users can also add default values and validity predicates.