Report errors while connecting two CAF nodes
Connecting two CAF nodes can fail for several reasons:
- unexpected TCP disconnects
- an application ID mismatch
- a CAF version mismatch

Currently, CAF reports all of the above errors as
`disconnect_during_handshake`. Masking the true error makes it
impossible for users to troubleshoot CAF applications or to respond to
errors programmatically. For example, an application may try to
reconnect on actual TCP disconnects but there's no point in
disconnecting to an incompatible node.