Add to_strings for enum back to the repository
Using a code generator that requires compiling a C++ tool breaks when
crosscompiling CAF. It can also interact badly with projects that
include CAF as a submodule.

As a compromise between easy consumption / integration of CAF and
maintainability of the `to_string` implementations, this set of changes
adds two new targets to CMake:
- `check-consistency` runs the code generator and checks whether all
  files in the repository are still up-to-date. We can run this target
  automatically in our CI to fail builds that introduce inconsistencies.
- `update-enum-strings` runs the code generator to replace the files in
  the source tree. This is meant for CAF developers to update `.cpp`
  files after changing enum headers.

Both new targets are excluded from `ALL` in CMake and thus do not run
automatically.