Remove the global `to_string` overload
The global `to_string` overload may have seemed like a good idea at the
time (4 years ago), but it interacts very badly with types such as
`variant` that have implicit conversion operators. The global
`to_string` also has a very weird interaction with the
`stringification_inspector`: we enable the `to_string` if a type is
inspectable but we use this as a fallback anyway. However, a custom
`to_string` that gets picked up via ADL can become ambiguous due to the
`caf::to_string` for no good reason.