Re-implement heartbeat using scheduled_send
By using `scheduled_send` instead of `delayed_send`, we can keep track
of the timeliness of our heartbeats. If the BASP broker processes its
tick after the next interval is over, we can assume an overload
situation or too low heartbeat intervals.

Further, the type of the `heartbeat_interval` constant was changed to
`timespan` in order to fix a mismatch between the actor system config
and the BASP broker. Relates #1095.