Simplify lifetime management of socket managers
Our previous design for the endpoint managers required us to use rely on
a shared ownership model for the multiplexer, because actor proxies held
a pointer to the socket manager directly. In our revised design, proxies
and managers are going to use a shared queue instead. This change
enables us to simplify the ownership model to a simple parent child
relation between the middleman and the multiplexer.