Check whether CMAKE_INSTALL_PREFIX is absolute
CAF uses `CMAKE_INSTALL_FULL_<GNUInstallDir>` instead of
`CMAKE_INSTALL_<GNUInstallDir>`, which by default are the latter
prefixed with `CMAKE_INSTALL_PREFIX`. However, there is no further
checking in place whether the resulting
`CMAKE_INSTALL_FULL_<GNUInstallDir>` is absolute, and if it is not, then
using it as a destination in `installl(... DESTINATION
${CMAKE_INSTALL_FULL_<GNUInstallDir>})` will cause it to be prefixed
with `CMAKE_INSTALL_PREFIX` again.