Implement new release semantics for flow handles
The subscription and the observer hold mutual references, which makes it
very easy to accidentally destroy the current object when disposing
objects. The `release_later` functions delay releasing the reference
count of discarded objects until all actions have completed. This makes
it safe to discard a subscription or observer, even if both objects end
up discard each other by delaying the destruction of the objects.