Fix exports and remove OBJECT libraries
We used the OBJECT libraries to give the tests the same access to
implementation details as the library under test. However, this
introduces multiple definitions of the same symbols in the test
binaries. For example, the core tests pull in all objects from
libcaf_core directly. However, the tests also depend on libcaf_test,
which pulls in libcaf_core as a dependency. The cleaner solution is to
just separate the test binary and the library target. This also revealed
several issues with missing (or wrong) exports.