Allow actors to send messages during cleanup
If the reference count of an actor reaches zero, it must still be able
to run its cleanup code. Since the reference count has already reached
zero, users had to be very careful to only do "safe" operations in
cleanup code. Most notably, this meant not sending messages.

This redesign tweaks the shutdown procedure as follows. When hitting 0,
CAF checks if an actor has not yet terminated. In case the actor must
still properly terminate, CAF increases the reference count back to 1
before calling the cleanup logic. Hence, the reference count will always
be at least 1 when running cleanup code and users no longer have to be
extra careful in cleanup code (such as the destructor of a state class).