Remove the behavior_changer mixin
Using mixin-based composition is great for prototyping. However, mixins
are not a first-class thing in C++ and the template-based mixins in CAF
make it harder to read the code for newcomers (and tools). Not using the
mixin for inserting `become()` and `unbecome()` introduces a bit of
code duplication, but overall makes the code base easier to follow.