Remove logger factory from the config class
The logger is a central component of the actor system. Allowing users to
override this component should be carefully considered. For now, we
should rather play it safe and not introduce a major feature by accident
that will make it hard to change central interfaces such as the logger
or the scheduler later on.