Deprecate the class-based spawn and actor pools
The class-based spawn overload is deprecated and users should instead
use the function-based and state-based spawn overloads.

The actor pool has a very cumbersome interface that is hard to use and
comes with several limitations. Eventually, we want to replace this
utility with shared mailboxes, rendering the pool-based approach
obsolete.