Always install a default logger when running tests
When using the deterministic fixture, we install a logger that forwards
all messages to the reporter. However, when not using the fixture, logs
are silently discarded because no logger exists. Since we offer a log
component specifically for testing, e.g. via `log::test::debug`, this is
not what a user can reasonably expect.

This patch moves the logger implementation from the deterministic
fixture to the reporter and always installs an instance of that logger
as default.