Make `from_container` work with `std::generator`
This fixes `from_container` to work with `std::generator`. Additionally,
it makes `from_generator` and `from_callable` work with non-copyable
callbacks.

I'm opening this pull request primarily for discussion, as this
currently breaks the `retry` flow operator when used in combination with
`from_container`, as that no longer restarts the iteration from the
start of the container. I am uncertain what to do about that. I do
personally think that the benefit of this change far outweighs its
downsides, but this should definitely be discussed.

Note that this is a breaking change for public-facing APIs, so _iff_
this change makes it should probably land in the `release/2` branch.

See also:
- Fixes #1968