Use consistent error codes for discarded requests
If an actor terminates, it will discard all messages from its mailbox.
Any discarded request message will receive an response message with an
error to inform the sender that the request has been discarded. The same
happens if a message has been discarded immediately because the mailbox
has already been closed.

Previously, CAF has used the exit reason passed to `quit`. However, this
makes it hard to differentiate between a discarded request and a request
that resulted in an error. Furthermore, this could result in error
messages with a default-constructed error code, which semantically makes
no sense.