Add new functions for operating on SPSC buffers
The new `consume_on` function on consumer resources enables actors to
consume items from a consumer resource without turning it into a flow.

The new `produce_on` function on producer resources is the symmetric
counterpart to `consume_on`.

These additions allow developers to get a more "direct" way to operate
on SPSC buffers whenever flows are not a good fit.