Deprecate boolean conversions for caf::error
Motivating example:

```c++
  if (try_fetch_data()) {
    // ...
  }
```

One would assume that the `if` branch executes if fetching was
successful. However, the opposite happens if the function has been
defined as:

```c++
caf::error try_fetch_data();
```

This makes it all too easy to introduce subtle bugs. Hence, we are
deprecating the boolean conversion in favor of requiring users to
explicitly call `empty()` or `valid()`.